+++ packages/manager-ui-kit/src/components/modal/Modal.component.tsx
@@ -1,4 +1,11 @@
-import { ElementRef, ForwardedRef, forwardRef, useMemo } from 'react';
+import {
+  ElementRef,
+  ForwardRefExoticComponent,
+  ForwardedRef,
+  RefAttributes,
+  forwardRef,
+  useMemo,
+} from 'react';
 
 import { useTranslation } from 'react-i18next';
 
@@ -19,7 +26,9 @@ import { NAMESPACES } from '@ovh-ux/manager-common-translations';
 
 import { ModalProps } from '@/components/modal/Modal.props';
 
-export const Modal = forwardRef(
+type ModalRef = ElementRef<typeof ModalContent>;
+
+export const Modal: ForwardRefExoticComponent<ModalProps & RefAttributes<ModalRef>> = forwardRef(
   (
     {
       heading,